requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
tqdm>=4.65.0
matplotlib>=3.8.0
datasets>=2.14.0
huggingface-hub>=0.16.0
torchinfo
